pyproject.toml
[project]
name = "t5-distractor-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.9.0",
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "huggingface-hub>=0.34.3",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.1.7",
    "kernels>=0.9.0",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "ollama>=0.5.3",
    "optimum-neuron>=0.0.3",
    "python-dotenv>=1.1.1",
    "rouge-score>=0.1.2",
    "sentencepiece>=0.2.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers==4.55.0",
    "wandb>=0.21.0",
]
